fix a bug for 2D and 3D data cases: when the original data size is not big while the precision is set to very high, the compression size may be larger than the original size thus leading to unexpected faults. (solution: use original data to replace the compressed data in this case - no compression mode instead)

+++ sz/include/DynamicDoubleArray.h
@@ -25,7 +25,7 @@ void new_DDA(DynamicDoubleArray **dda, int cap);
 void convertDDAtoDoubles(DynamicDoubleArray *dba, double **data);
 void free_DDA(DynamicDoubleArray *dda);
 double getDDA_Data(DynamicDoubleArray *dda, int pos);
-void addDDA_Data(DynamicDoubleArray *dda, int value);
+void addDDA_Data(DynamicDoubleArray *dda, double value);
 
 #ifdef __cplusplus
 }
